==> python/sample/20-get-census-data.ipynb <==
import pandas as pd
import urllib2 
import json
import csv 

acsLoc = '/Volumes/Echo/GIS/projects/naas/data/input/census/tables/acs5yr/2012/'
cenLoc = '/Volumes/Echo/GIS/projects/naas/data/input/census/tables/census/2000/'

print 'download json to csv for pop'
topicnm = 'pop'
url = 'http://api.census.gov/data/2012/acs5?key=30699f15ab4d04a1e0943715b539d256c9a3ee44&get=B01003_001E&for=tract:*&in=state:36'
data = urllib2.urlopen(url).read()
data = json.loads(data)
fname = acsLoc+"tract_st36_"+topicnm+".csv"
with open(fname,'wb') as outf:
    outcsv = csv.writer(outf)
    outcsv.writerows(data)

url = 'http://api.census.gov/data/2000/sf3?key=30699f15ab4d04a1e0943715b539d256c9a3ee44&get=P001001&for=tract:*&in=state:36'
data = urllib2.urlopen(url).read()
data = json.loads(data)
fname = cenLoc+"tract_st36_"+topicnm+".csv"
with open(fname,'wb') as outf:
    outcsv = csv.writer(outf)
    outcsv.writerows(data)
    
print 'download json to csv for pov'    
topicnm = 'pov'  
fields = 'C17002_001E,C17002_002E,C17002_003E'
url = 'http://api.census.gov/data/2012/acs5?key=30699f15ab4d04a1e0943715b539d256c9a3ee44&get='+fields+'&for=tract:*&in=state:36'
data = urllib2.urlopen(url).read()
data = json.loads(data)
fname = acsLoc+"tract_st36_"+topicnm+".csv"
with open(fname,'wb') as outf:
    outcsv = csv.writer(outf)
    outcsv.writerows(data)

fields = 'P088001,P088002,P088003,P088004,P088005,P088006,P088007,P088008,P088009,P088010'
url = 'http://api.census.gov/data/2000/sf3?key=30699f15ab4d04a1e0943715b539d256c9a3ee44&get='+fields+'&for=tract:*&in=state:36'
data = urllib2.urlopen(url).read()
data = json.loads(data)
fname = cenLoc+"tract_st36_"+topicnm+".csv"
with open(fname,'wb') as outf:
    outcsv = csv.writer(outf)
    outcsv.writerows(data)   
    
print 'download json to csv for mhi'    
topicnm = 'mhi'   
fields = 'B19013_001E'
url = 'http://api.census.gov/data/2012/acs5?key=30699f15ab4d04a1e0943715b539d256c9a3ee44&get='+fields+'&for=tract:*&in=state:36'
data = urllib2.urlopen(url).read()
data = json.loads(data)
fname = acsLoc+"tract_st36_"+topicnm+".csv"
with open(fname,'wb') as outf:
    outcsv = csv.writer(outf)
    outcsv.writerows(data)
    
fields = 'P053001'
url = 'http://api.census.gov/data/2000/sf3?key=30699f15ab4d04a1e0943715b539d256c9a3ee44&get='+fields+'&for=tract:*&in=state:36'
data = urllib2.urlopen(url).read()
data = json.loads(data)
fname = cenLoc+"tract_st36_"+topicnm+".csv"
with open(fname,'wb') as outf:
    outcsv = csv.writer(outf)
    outcsv.writerows(data)


import pandas as pd
import numpy as np

acsLoc = '/Volumes/Echo/GIS/projects/naas/data/input/census/tables/acs5yr/2012/'
cenLoc = '/Volumes/Echo/GIS/projects/naas/data/input/census/tables/census/2000/'

pov00 = pd.read_csv(cenLoc+'tract_st36_pov.csv')
pop00 = pd.read_csv(cenLoc+'tract_st36_pop.csv')
mhi00 = pd.read_csv(cenLoc+'tract_st36_mhi.csv')
pov10 = pd.read_csv(acsLoc+'tract_st36_pov.csv')
pop10 = pd.read_csv(acsLoc+'tract_st36_pop.csv')
mhi10 = pd.read_csv(acsLoc+'tract_st36_mhi.csv')

tList = [pov00,pop00,mhi00,pov10,pop10,mhi10]

for df in tList:
    df['state'] = df['state'].astype(str)
    df['county'] = df['county'].astype(str)
    df['tract'] = df['tract'].astype(str)
    
    df['county'] = df['county'].apply(lambda x: '{0:0>3}'.format(x)) #add leaving zeros
    df['tract'] = df['tract'].apply(lambda x: '{0:0>6}'.format(x))
    df = df.rename(columns=lambda x: x.replace('_', ''))
    
pov00['geoid'] = pov00['state'] + pov00['county'] + pov00['tract']
pop00['geoid'] = pop00['state'] + pop00['county'] + pop00['tract']
mhi00['geoid'] = mhi00['state'] + mhi00['county'] + mhi00['tract']
pov10['geoid'] = pov10['state'] + pov10['county'] + pov10['tract']
pop10['geoid'] = pop10['state'] + pop10['county'] + pop10['tract']
mhi10['geoid'] = mhi10['state'] + mhi10['county'] + mhi10['tract']

pov00 = pov00.drop(['state','county','tract'], axis=1)
pop00 = pop00.drop(['state','county','tract'], axis=1)
mhi00 = mhi00.drop(['state','county','tract'], axis=1)
pov10 = pov10.drop(['state','county','tract'], axis=1)
pop10 = pop10.drop(['state','county','tract'], axis=1)
mhi10 = mhi10.drop(['state','county','tract'], axis=1)

pov00.to_csv(cenLoc+'tract_st36_pov_cln.csv',index=False)
pop00.to_csv(cenLoc+'tract_st36_pop_cln.csv',index=False)
mhi00.to_csv(cenLoc+'tract_st36_mhi_cln.csv',index=False)
pov10.to_csv(acsLoc+'tract_st36_pov_cln.csv',index=False)
pop10.to_csv(acsLoc+'tract_st36_pop_cln.csv',index=False)
mhi10.to_csv(acsLoc+'tract_st36_mhi_cln.csv',index=False)

pov00.head(10)

List of Census Variables

* Population density (2000 and 2010 census) - thinkign ACS as I can get financial data
* % poverty (2000 and 2010 census)

    * 2012 acs5yr
    
            **^^^pctpov
            Percent population in poverty in Neighborhood Geography (**).
            American Community Survey 2008-2012 (5-year)
            ( df[geo+'C17002002E'] + df[geo+'C17002003E'] ) / df[geo+'C17002001E']

    * 2000 census sf3
        
             RATIO OF INCOME IN 1999 TO POVERTY LEVEL
            [10]
            Universe: Population for whom poverty status is
            determined
            Total: P088001 07 9
            Under .50 P088002 07 9
            .50 to .74 P088003 07 9
            .75 to .99 P088004 07 9
            1.00 to 1.24 P088005 07 9
            1.25 to 1.49 P088006 07 9
            1.50 to 1.74 P088007 07 9
            1.75 to 1.84 P088008 07 9
            1.85 to 1.99 P088009 07 9
            2.00 and over P088010 07 9
            
* Median household income (2000 and 2010 census)
    * 2012 acs5yr
    
            **^^^medhhinc
            Median Household Income in Neighborhood Geography (**).
            American Community Survey 2008-2012 (5-year)
            df[geo+'B19013001E'] / df[geo+'countrows']
            
    * 2000 census sf3
            
            P53. MEDIAN HOUSEHOLD INCOME IN 1999
            (DOLLARS) [1]
            Universe: Households
            Median household income in 1999 P053001 06 9
