==> python/sample/21-census-tables.ipynb <==
import pandas as pd
geogList     = ['h','s']
radbufListFn = ['r0050m','r0100m','r0250m','r0500m','r1000m']

for j in geogList:
    for k in radbufListFn:
        df00 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis.csv')
        df10 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis.csv')
        df00 = df00[['UID','GEOID','CT2000AREA','NEWAREA']]
        df10 = df10[['UID','GEOID','CT2010AREA','NEWAREA']]
        df00 = df00.rename(columns=lambda x: x.lower())
        df10 = df10.rename(columns=lambda x: x.lower())
        df00['pctorigarea'] = df00['newarea'] / df00['ct2000area']
        df10['pctorigarea'] = df10['newarea'] / df10['ct2010area']
        df00['uid'] = df00['uid'].astype(int)
        df10['uid'] = df10['uid'].astype(int)
        df00 = df00[['uid','geoid','pctorigarea','newarea']]
        df10 = df10[['uid','geoid','pctorigarea','newarea']]
        df00.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln.csv', index=False)
        df10.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln.csv', index=False)
    


import pandas as pd
geogList     = ['h','s']
netbufListFn = ['n1000m']

for j in geogList:
    for k in netbufListFn:
        df00 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis.csv')
        df10 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis.csv')
        df00 = df00[['UID','GEOID','CT2000AREA','NEWAREA']]
        df10 = df10[['UID','GEOID','CT2010AREA','NEWAREA']]
        df00 = df00.rename(columns=lambda x: x.lower())
        df10 = df10.rename(columns=lambda x: x.lower())
        df00['pctorigarea'] = df00['newarea'] / df00['ct2000area']
        df10['pctorigarea'] = df10['newarea'] / df10['ct2010area']
        df00['uid'] = df00['uid'].astype(int)
        df10['uid'] = df10['uid'].astype(int)
        df00 = df00[['uid','geoid','pctorigarea','newarea']]
        df10 = df10[['uid','geoid','pctorigarea','newarea']]
        df00.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln.csv', index=False)
        df10.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln.csv', index=False)
    

import pandas as pd
geogList     = ['h','s']
radbufListFn = ['r0050m','r0100m','r0250m','r0500m','r1000m']

acsLoc = '/Volumes/Echo/GIS/projects/naas/data/input/census/tables/acs5yr/2012/'
cenLoc = '/Volumes/Echo/GIS/projects/naas/data/input/census/tables/census/2000/'

pov00 = pd.read_csv(cenLoc+'tract_st36_pov_cln.csv')
pop00 = pd.read_csv(cenLoc+'tract_st36_pop_cln.csv')
mhi00 = pd.read_csv(cenLoc+'tract_st36_mhi_cln.csv')
pov10 = pd.read_csv(acsLoc+'tract_st36_pov_cln.csv')
pop10 = pd.read_csv(acsLoc+'tract_st36_pop_cln.csv')
mhi10 = pd.read_csv(acsLoc+'tract_st36_mhi_cln.csv')

ct00 = pop00.merge(pov00, on='geoid').merge(mhi00, on='geoid')
ct10 = pop10.merge(pov10, on='geoid').merge(mhi10, on='geoid')

tList = [pov00,pop00,mhi00,pov10,pop10,mhi10]

for j in geogList:
    for k in radbufListFn:
        df00 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln.csv')
        df10 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln.csv')
            
        df00 = df00.merge(ct00, on='geoid', how='left')
        df10 = df10.merge(ct10, on='geoid', how='left')
        
        df00.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln_vars.csv', index=False)
        df10.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln_vars.csv', index=False)
        
        #print df00.head(5)

import pandas as pd
geogList     = ['h','s']
netbufListFn = ['n1000m']

acsLoc = '/Volumes/Echo/GIS/projects/naas/data/input/census/tables/acs5yr/2012/'
cenLoc = '/Volumes/Echo/GIS/projects/naas/data/input/census/tables/census/2000/'

pov00 = pd.read_csv(cenLoc+'tract_st36_pov_cln.csv')
pop00 = pd.read_csv(cenLoc+'tract_st36_pop_cln.csv')
mhi00 = pd.read_csv(cenLoc+'tract_st36_mhi_cln.csv')
pov10 = pd.read_csv(acsLoc+'tract_st36_pov_cln.csv')
pop10 = pd.read_csv(acsLoc+'tract_st36_pop_cln.csv')
mhi10 = pd.read_csv(acsLoc+'tract_st36_mhi_cln.csv')

ct00 = pop00.merge(pov00, on='geoid').merge(mhi00, on='geoid')
ct10 = pop10.merge(pov10, on='geoid').merge(mhi10, on='geoid')

tList = [pov00,pop00,mhi00,pov10,pop10,mhi10]

for j in geogList:
    for k in netbufListFn:
        df00 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln.csv')
        df10 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln.csv')
            
        df00 = df00.merge(ct00, on='geoid', how='left')
        df10 = df10.merge(ct10, on='geoid', how='left')
        
        df00.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln_vars.csv', index=False)
        df10.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln_vars.csv', index=False)
        
        #print df00.head(5)

#For Population and Poverty
import pandas as pd
geogList     = ['h','s']
radbufListFn = ['r0050m','r0100m','r0250m','r0500m','r1000m']

vars00 = ['P001001','P088001','P088002','P088003','P088004']
vNam00 = ['cen00pop','cen00pov1','cen00pov2','cen00pov3','cen00pov4']
vars10 = ['B01003_001E','C17002_001E','C17002_002E','C17002_003E']
vNam10 = ['acs12pop','acs12pov1','acs12pov2','acs12pov3']

for j in geogList:
    for k in radbufListFn:
        df00 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln_vars.csv')
        df10 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln_vars.csv')
        
        for h,i in zip(vars00,vNam00):
            df00[j+k+i] = df00[h] * df00['pctorigarea']
            
        for h,i in zip(vars10,vNam10):
            df10[j+k+i] = df10[h] * df10['pctorigarea']      
        
        df00 = df00.drop(['P001001','P088001','P088002','P088003','P088004','P088005','P088006','P088007','P088008','P088009','P088010','P053001'], axis=1)
        df10 = df10.drop(['B01003_001E','C17002_001E','C17002_002E','C17002_003E','B19013_001E'], axis=1)
        
        dfg00 = df00.groupby(['uid']).sum()
        dfg10 = df10.groupby(['uid']).sum()
        
        dfg00 = dfg00.drop(['geoid','pctorigarea'], axis=1)
        dfg10 = dfg10.drop(['geoid','pctorigarea'], axis=1)
        
        dfg00[j+k+'cen00pctpov'] = ( dfg00[j+k+'cen00pov2'] + dfg00[j+k+'cen00pov3'] + dfg00[j+k+'cen00pov4'] )  / dfg00[j+k+'cen00pov1']
        dfg10[j+k+'acs12pctpov'] = ( dfg10[j+k+'acs12pov2'] + dfg10[j+k+'acs12pov3'] )  / dfg10[j+k+'acs12pov1']
        
        dfg00.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln_vars_mult_grpby.csv')
        dfg10.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln_vars_mult_grpby.csv')
        

#For Population and Poverty
import pandas as pd
geogList     = ['h','s']
netbufListFn = ['n1000m']

vars00 = ['P001001','P088001','P088002','P088003','P088004']
vNam00 = ['cen00pop','cen00pov1','cen00pov2','cen00pov3','cen00pov4']
vars10 = ['B01003_001E','C17002_001E','C17002_002E','C17002_003E']
vNam10 = ['acs12pop','acs12pov1','acs12pov2','acs12pov3']

for j in geogList:
    for k in netbufListFn:
        df00 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln_vars.csv')
        df10 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln_vars.csv')
        
        for h,i in zip(vars00,vNam00):
            df00[j+k+i] = df00[h] * df00['pctorigarea']
            
        for h,i in zip(vars10,vNam10):
            df10[j+k+i] = df10[h] * df10['pctorigarea']      
        
        df00 = df00.drop(['P001001','P088001','P088002','P088003','P088004','P088005','P088006','P088007','P088008','P088009','P088010','P053001'], axis=1)
        df10 = df10.drop(['B01003_001E','C17002_001E','C17002_002E','C17002_003E','B19013_001E'], axis=1)
        
        dfg00 = df00.groupby(['uid']).sum()
        dfg10 = df10.groupby(['uid']).sum()
        
        dfg00 = dfg00.drop(['geoid','pctorigarea'], axis=1)
        dfg10 = dfg10.drop(['geoid','pctorigarea'], axis=1)
        
        dfg00[j+k+'cen00pctpov'] = ( dfg00[j+k+'cen00pov2'] + dfg00[j+k+'cen00pov3'] + dfg00[j+k+'cen00pov4'] )  / dfg00[j+k+'cen00pov1']
        dfg10[j+k+'acs12pctpov'] = ( dfg10[j+k+'acs12pov2'] + dfg10[j+k+'acs12pov3'] )  / dfg10[j+k+'acs12pov1']
        
        dfg00.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln_vars_mult_grpby.csv')
        dfg10.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln_vars_mult_grpby.csv')


#HOME For Population and Poverty
o = 'outer'
u = 'uid'

tm =  '/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/'

df = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/input/addresses/geocode_success_home.csv')
df = df[['uid']]

df1 = pd.read_csv(tm+'census/hr0050m_land_int_ct00_dis_cln_vars_mult_grpby.csv')
df2 = pd.read_csv(tm+'census/hr0100m_land_int_ct00_dis_cln_vars_mult_grpby.csv')
df3 = pd.read_csv(tm+'census/hr0250m_land_int_ct00_dis_cln_vars_mult_grpby.csv')
df4 = pd.read_csv(tm+'census/hr0500m_land_int_ct00_dis_cln_vars_mult_grpby.csv')
df5 = pd.read_csv(tm+'census/hr1000m_land_int_ct00_dis_cln_vars_mult_grpby.csv')

df6 = pd.read_csv(tm+'census/hr0050m_land_int_ct10_dis_cln_vars_mult_grpby.csv')
df7 = pd.read_csv(tm+'census/hr0100m_land_int_ct10_dis_cln_vars_mult_grpby.csv')
df8 = pd.read_csv(tm+'census/hr0250m_land_int_ct10_dis_cln_vars_mult_grpby.csv')
df9 = pd.read_csv(tm+'census/hr0500m_land_int_ct10_dis_cln_vars_mult_grpby.csv')
df10= pd.read_csv(tm+'census/hr1000m_land_int_ct10_dis_cln_vars_mult_grpby.csv')

df21= pd.read_csv(tm+'census/hn1000m_land_int_ct00_dis_cln_vars_mult_grpby.csv').drop('newarea',axis=1)

df26= pd.read_csv(tm+'census/hn1000m_land_int_ct10_dis_cln_vars_mult_grpby.csv').drop('newarea',axis=1)


df = df.merge(df1, on=u, how=o).merge(df2, on=u, how=o).merge(df3, on=u, how=o).merge(df4, on=u, how=o).merge(df5, on=u, how=o)
df = df.merge(df6, on=u, how=o).merge(df7, on=u, how=o).merge(df8, on=u, how=o).merge(df9, on=u, how=o).merge(df10, on=u, how=o)
df = df.merge(df21, on=u, how=o).merge(df26, on=u, how=o)

df.to_csv(tm+'census/h_census_poppov.csv', index=False)

#SCHOOL For Population and Poverty
o = 'outer'
u = 'uid'

tm =  '/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/'

df = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/input/addresses/geocode_success_scho.csv')
df = df[['uid']]

df1 = pd.read_csv(tm+'census/sr0050m_land_int_ct00_dis_cln_vars_mult_grpby.csv')
df2 = pd.read_csv(tm+'census/sr0100m_land_int_ct00_dis_cln_vars_mult_grpby.csv')
df3 = pd.read_csv(tm+'census/sr0250m_land_int_ct00_dis_cln_vars_mult_grpby.csv')
df4 = pd.read_csv(tm+'census/sr0500m_land_int_ct00_dis_cln_vars_mult_grpby.csv')
df5 = pd.read_csv(tm+'census/sr1000m_land_int_ct00_dis_cln_vars_mult_grpby.csv')

df6 = pd.read_csv(tm+'census/sr0050m_land_int_ct10_dis_cln_vars_mult_grpby.csv')
df7 = pd.read_csv(tm+'census/sr0100m_land_int_ct10_dis_cln_vars_mult_grpby.csv')
df8 = pd.read_csv(tm+'census/sr0250m_land_int_ct10_dis_cln_vars_mult_grpby.csv')
df9 = pd.read_csv(tm+'census/sr0500m_land_int_ct10_dis_cln_vars_mult_grpby.csv')
df10= pd.read_csv(tm+'census/sr1000m_land_int_ct10_dis_cln_vars_mult_grpby.csv')

df21= pd.read_csv(tm+'census/sn1000m_land_int_ct00_dis_cln_vars_mult_grpby.csv').drop('newarea',axis=1)

df26= pd.read_csv(tm+'census/sn1000m_land_int_ct10_dis_cln_vars_mult_grpby.csv').drop('newarea',axis=1)


df = df.merge(df1, on=u, how=o).merge(df2, on=u, how=o).merge(df3, on=u, how=o).merge(df4, on=u, how=o).merge(df5, on=u, how=o)
df = df.merge(df6, on=u, how=o).merge(df7, on=u, how=o).merge(df8, on=u, how=o).merge(df9, on=u, how=o).merge(df10, on=u, how=o)
df = df.merge(df21, on=u, how=o).merge(df26, on=u, how=o)

df.to_csv(tm+'census/s_census_poppov.csv', index=False)

#For Median Household Income
import pandas as pd
geogList     = ['h','s']
radbufListFn = ['r0050m','r0100m','r0250m','r0500m','r1000m']

for j in geogList:
    for k in radbufListFn:
        df00 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln_vars.csv')
        df10 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln_vars.csv')
        
        df00 = df00[['uid','geoid','pctorigarea','newarea','P053001']]
        df10 = df10[['uid','geoid','pctorigarea','newarea','B19013_001E']]
        
        df00 = df00[( df00.P053001 > 0 )]
        df10 = df10[( df10.B19013_001E > 0 )]
        
        df00gby = df00.groupby('uid').sum()
        df10gby = df10.groupby('uid').sum()
        
        df00gby.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct00_area_for_mhhi.csv')
        df10gby.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct10_area_for_mhhi.csv')
        
        df00area = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct00_area_for_mhhi.csv')
        df10area = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct10_area_for_mhhi.csv')
        
        df00area = df00area[['uid','newarea']]
        df10area = df10area[['uid','newarea']]
        
        df00area[j+k+'cen00mhhi_totarea'] = df00area['newarea']
        df10area[j+k+'acs12mhhi_totarea'] = df10area['newarea']
          
        df00area = df00area[['uid',j+k+'cen00mhhi_totarea']]
        df10area = df10area[['uid',j+k+'acs12mhhi_totarea']]      
                 
        df00area.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct00_area_for_mhhi.csv')
        df10area.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct10_area_for_mhhi.csv')
        
        df00 = df00.merge(df00area, on='uid', how='left')
        df10 = df10.merge(df10area, on='uid', how='left')
        
        df00['pcttotmmhiarea'] = df00['newarea'] / df00[j+k+'cen00mhhi_totarea'] 
        df10['pcttotmmhiarea'] = df10['newarea'] / df10[j+k+'acs12mhhi_totarea'] 
        
        df00[j+k+'cen10mhhi'] = df00['pcttotmmhiarea'] * df00['P053001']
        df10[j+k+'acs12mhhi'] = df10['pcttotmmhiarea'] * df10['B19013_001E']
        
        df00 = df00[['uid',j+k+'cen10mhhi']]
        df10 = df10[['uid',j+k+'acs12mhhi']]
        
        df00gbymmhi = df00.groupby('uid').sum()
        df10gbymmhi = df10.groupby('uid').sum()
        
        df00gbymmhi.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct00_area_for_mhhi_groupby_final.csv')
        df10gbymmhi.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct10_area_for_mhhi_groupby_final.csv')
        
        print df00gbymmhi
        
        

#For Median Household Income
import pandas as pd
geogList     = ['h','s']
netbufListFn = ['n1000m']

for j in geogList:
    for k in netbufListFn:
        df00 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct00_dis_cln_vars.csv')
        df10 = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_land_int_ct10_dis_cln_vars.csv')
        
        df00 = df00[['uid','geoid','pctorigarea','newarea','P053001']]
        df10 = df10[['uid','geoid','pctorigarea','newarea','B19013_001E']]
        
        df00 = df00[( df00.P053001 > 0 )]
        df10 = df10[( df10.B19013_001E > 0 )]
        
        df00gby = df00.groupby('uid').sum()
        df10gby = df10.groupby('uid').sum()
        
        df00gby.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct00_area_for_mhhi.csv')
        df10gby.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct10_area_for_mhhi.csv')
        
        df00area = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct00_area_for_mhhi.csv')
        df10area = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct10_area_for_mhhi.csv')
        
        df00area = df00area[['uid','newarea']]
        df10area = df10area[['uid','newarea']]
        
        df00area[j+k+'cen00mhhi_totarea'] = df00area['newarea']
        df10area[j+k+'acs12mhhi_totarea'] = df10area['newarea']
          
        df00area = df00area[['uid',j+k+'cen00mhhi_totarea']]
        df10area = df10area[['uid',j+k+'acs12mhhi_totarea']]      
                 
        df00area.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct00_area_for_mhhi.csv')
        df10area.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct10_area_for_mhhi.csv')
        
        df00 = df00.merge(df00area, on='uid', how='left')
        df10 = df10.merge(df10area, on='uid', how='left')
        
        df00['pcttotmmhiarea'] = df00['newarea'] / df00[j+k+'cen00mhhi_totarea'] 
        df10['pcttotmmhiarea'] = df10['newarea'] / df10[j+k+'acs12mhhi_totarea'] 
        
        df00[j+k+'cen10mhhi'] = df00['pcttotmmhiarea'] * df00['P053001']
        df10[j+k+'acs12mhhi'] = df10['pcttotmmhiarea'] * df10['B19013_001E']
        
        df00 = df00[['uid',j+k+'cen10mhhi']]
        df10 = df10[['uid',j+k+'acs12mhhi']]
        
        df00gbymmhi = df00.groupby('uid').sum()
        df10gbymmhi = df10.groupby('uid').sum()
        
        df00gbymmhi.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct00_area_for_mhhi_groupby_final.csv')
        df10gbymmhi.to_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/census/'+j+k+'_ct10_area_for_mhhi_groupby_final.csv')
        
        print df00gbymmhi
        
        

#HOME For MHHI
o = 'outer'
u = 'uid'

tm =  '/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/'

df = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/input/addresses/geocode_success_home.csv')
df = df[['uid']]

df1 = pd.read_csv(tm+'census/hr0050m_ct00_area_for_mhhi_groupby_final.csv')
df2 = pd.read_csv(tm+'census/hr0100m_ct00_area_for_mhhi_groupby_final.csv')
df3 = pd.read_csv(tm+'census/hr0250m_ct00_area_for_mhhi_groupby_final.csv')
df4 = pd.read_csv(tm+'census/hr0500m_ct00_area_for_mhhi_groupby_final.csv')
df5 = pd.read_csv(tm+'census/hr1000m_ct00_area_for_mhhi_groupby_final.csv')

df6 = pd.read_csv(tm+'census/hr0050m_ct10_area_for_mhhi_groupby_final.csv')
df7 = pd.read_csv(tm+'census/hr0100m_ct10_area_for_mhhi_groupby_final.csv')
df8 = pd.read_csv(tm+'census/hr0250m_ct10_area_for_mhhi_groupby_final.csv')
df9 = pd.read_csv(tm+'census/hr0500m_ct10_area_for_mhhi_groupby_final.csv')
df10= pd.read_csv(tm+'census/hr1000m_ct10_area_for_mhhi_groupby_final.csv')

df21= pd.read_csv(tm+'census/hn1000m_ct00_area_for_mhhi_groupby_final.csv')

df26= pd.read_csv(tm+'census/hn1000m_ct10_area_for_mhhi_groupby_final.csv')

df = df.merge(df1, on=u, how=o).merge(df2, on=u, how=o).merge(df3, on=u, how=o).merge(df4, on=u, how=o).merge(df5, on=u, how=o)
df = df.merge(df6, on=u, how=o).merge(df7, on=u, how=o).merge(df8, on=u, how=o).merge(df9, on=u, how=o).merge(df10, on=u, how=o)
df = df.merge(df21, on=u, how=o).merge(df26, on=u, how=o)

df.to_csv(tm+'census/h_census_mhhi.csv', index=False)

#SCHOOL For MHHI
o = 'outer'
u = 'uid'

tm =  '/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/tables/'

df = pd.read_csv('/Volumes/Echo/GIS/projects/naas/tasks/201502_naas_locations/data/input/addresses/geocode_success_scho.csv')
df = df[['uid']]

df1 = pd.read_csv(tm+'census/sr0050m_ct00_area_for_mhhi_groupby_final.csv')
df2 = pd.read_csv(tm+'census/sr0100m_ct00_area_for_mhhi_groupby_final.csv')
df3 = pd.read_csv(tm+'census/sr0250m_ct00_area_for_mhhi_groupby_final.csv')
df4 = pd.read_csv(tm+'census/sr0500m_ct00_area_for_mhhi_groupby_final.csv')
df5 = pd.read_csv(tm+'census/sr1000m_ct00_area_for_mhhi_groupby_final.csv')

df6 = pd.read_csv(tm+'census/sr0050m_ct10_area_for_mhhi_groupby_final.csv')
df7 = pd.read_csv(tm+'census/sr0100m_ct10_area_for_mhhi_groupby_final.csv')
df8 = pd.read_csv(tm+'census/sr0250m_ct10_area_for_mhhi_groupby_final.csv')
df9 = pd.read_csv(tm+'census/sr0500m_ct10_area_for_mhhi_groupby_final.csv')
df10= pd.read_csv(tm+'census/sr1000m_ct10_area_for_mhhi_groupby_final.csv')

df21= pd.read_csv(tm+'census/sn1000m_ct00_area_for_mhhi_groupby_final.csv')

df26= pd.read_csv(tm+'census/sn1000m_ct10_area_for_mhhi_groupby_final.csv')

df = df.merge(df1, on=u, how=o).merge(df2, on=u, how=o).merge(df3, on=u, how=o).merge(df4, on=u, how=o).merge(df5, on=u, how=o)
df = df.merge(df6, on=u, how=o).merge(df7, on=u, how=o).merge(df8, on=u, how=o).merge(df9, on=u, how=o).merge(df10, on=u, how=o)
df = df.merge(df21, on=u, how=o).merge(df26, on=u, how=o)

df.to_csv(tm+'census/s_census_mhhi.csv', index=False)
